+++ system/hardware/base.py
@@ -6,6 +6,19 @@ from cereal import log
 
 NetworkType = log.DeviceState.NetworkType
 
+class LPAError(RuntimeError):
+  pass
+
+class LPAProfileNotFoundError(LPAError):
+  pass
+
+@dataclass
+class Profile:
+  iccid: str
+  nickname: str
+  enabled: bool
+  provider: str
+
 @dataclass
 class ThermalZone:
   # a zone from /sys/class/thermal/thermal_zone*
@@ -50,6 +63,31 @@ class ThermalConfig:
         else:
           ret[f.name + "TempC"] = v.read()
     return ret
+
+class LPABase(ABC):
+  @abstractmethod
+  def list_profiles(self) -> list[Profile]:
+    pass
+
+  @abstractmethod
+  def get_active_profile(self) -> Profile | None:
+    pass
+
+  @abstractmethod
+  def delete_profile(self, iccid: str) -> None:
+    pass
+
+  @abstractmethod
+  def download_profile(self, qr: str, nickname: str | None = None) -> None:
+    pass
+
+  @abstractmethod
+  def nickname_profile(self, iccid: str, nickname: str) -> None:
+    pass
+
+  @abstractmethod
+  def switch_profile(self, iccid: str) -> None:
+    pass
 
 class HardwareBase(ABC):
   @staticmethod
@@ -103,6 +141,10 @@ class HardwareBase(ABC):
 
   @abstractmethod
   def get_sim_info(self):
+    pass
+
+  @abstractmethod
+  def get_sim_lpa(self) -> LPABase:
     pass
 
   @abstractmethod
