+++ system/hardware/pc/hardware.py
@@ -1,11 +1,10 @@
 import random
 
 from cereal import log
-from openpilot.system.hardware.base import HardwareBase
+from openpilot.system.hardware.base import HardwareBase, LPABase
 
 NetworkType = log.DeviceState.NetworkType
 NetworkStrength = log.DeviceState.NetworkStrength
-
 
 class Pc(HardwareBase):
   def get_os_version(self):
@@ -40,6 +39,9 @@ class Pc(HardwareBase):
       'sim_state': ["ABSENT"],
       'data_connected': False
     }
+
+  def get_sim_lpa(self) -> LPABase:
+    raise NotImplementedError("SIM LPA not implemented for PC")
 
   def get_network_strength(self, network_type):
     return NetworkStrength.unknown
