+++ system/hardware/tici/esim.py
@@ -1,28 +1,12 @@
-#!/usr/bin/env python3
-
-import argparse
 import json
 import os
 import shutil
 import subprocess
-from dataclasses import dataclass
 from typing import Literal
 
-@dataclass
-class Profile:
-  iccid: str
-  nickname: str
-  enabled: bool
-  provider: str
+from openpilot.system.hardware.base import LPABase, LPAError, LPAProfileNotFoundError, Profile
 
-class LPAError(RuntimeError):
-  pass
-
-class LPAProfileNotFoundError(LPAError):
-  pass
-
-
-class LPA:
+class TiciLPA(LPABase):
   def __init__(self, interface: Literal['qmi', 'at'] = 'qmi'):
     self.env = os.environ.copy()
     self.env['LPAC_APDU'] = interface
@@ -47,31 +31,13 @@ class LPA:
   def get_active_profile(self) -> Profile | None:
     return next((p for p in self.list_profiles() if p.enabled), None)
 
-  def enable_profile(self, iccid: str) -> None:
-    self._validate_profile_exists(iccid)
-    latest = self.get_active_profile()
-    if latest:
-      if latest.iccid == iccid:
-        return
-      self.disable_profile(latest.iccid)
-    self._validate_successful(self._invoke('profile', 'enable', iccid))
-    self.process_notifications()
-
-  def disable_profile(self, iccid: str) -> None:
-    self._validate_profile_exists(iccid)
-    latest = self.get_active_profile()
-    if latest is not None and latest.iccid != iccid:
-      return
-    self._validate_successful(self._invoke('profile', 'disable', iccid))
-    self.process_notifications()
-
   def delete_profile(self, iccid: str) -> None:
     self._validate_profile_exists(iccid)
     latest = self.get_active_profile()
     if latest is not None and latest.iccid == iccid:
-      self.disable_profile(iccid)
+      raise LPAError('cannot delete active profile, switch to another profile first')
     self._validate_successful(self._invoke('profile', 'delete', iccid))
-    self.process_notifications()
+    self._process_notifications()
 
   def download_profile(self, qr: str, nickname: str | None = None) -> None:
     msgs = self._invoke('profile', 'download', '-a', qr)
@@ -81,17 +47,24 @@ class LPA:
       raise LPAError('no new profile found')
     if nickname:
       self.nickname_profile(new_profile['payload']['data']['iccid'], nickname)
-    self.process_notifications()
+    self._process_notifications()
 
   def nickname_profile(self, iccid: str, nickname: str) -> None:
     self._validate_profile_exists(iccid)
     self._validate_successful(self._invoke('profile', 'nickname', iccid, nickname))
 
-  def process_notifications(self) -> None:
-    """
-    Process notifications stored on the eUICC, typically to activate/deactivate the profile with the carrier.
-    """
-    self._validate_successful(self._invoke('notification', 'process', '-a', '-r'))
+  def switch_profile(self, iccid: str) -> None:
+    self._enable_profile(iccid)
+
+  def _enable_profile(self, iccid: str) -> None:
+    self._validate_profile_exists(iccid)
+    latest = self.get_active_profile()
+    if latest:
+      if latest.iccid == iccid:
+        return
+      self._validate_successful(self._invoke('profile', 'disable', latest.iccid))
+    self._validate_successful(self._invoke('profile', 'enable', iccid))
+    self._process_notifications()
 
   def _invoke(self, *cmd: str):
     proc = subprocess.Popen(['sudo', '-E', 'lpac'] + list(cmd), stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=self.env)
@@ -124,47 +97,15 @@ class LPA:
 
     return messages
 
+  def _process_notifications(self) -> None:
+    """
+    Process notifications stored on the eUICC, typically to activate/deactivate the profile with the carrier.
+    """
+    self._validate_successful(self._invoke('notification', 'process', '-a', '-r'))
+
   def _validate_profile_exists(self, iccid: str) -> None:
     if not any(p.iccid == iccid for p in self.list_profiles()):
       raise LPAProfileNotFoundError(f'profile {iccid} does not exist')
 
   def _validate_successful(self, msgs: list[dict]) -> None:
     assert msgs[-1]['payload']['message'] == 'success', 'expected success notification'
-
-
-if __name__ == "__main__":
-  parser = argparse.ArgumentParser(prog='esim.py', description='manage eSIM profiles on your comma device', epilog='comma.ai')
-  parser.add_argument('--backend', choices=['qmi', 'at'], default='qmi', help='use the specified backend, defaults to qmi')
-  parser.add_argument('--enable', metavar='iccid', help='enable profile; will disable current profile')
-  parser.add_argument('--disable', metavar='iccid', help='disable profile')
-  parser.add_argument('--delete', metavar='iccid', help='delete profile (warning: this cannot be undone)')
-  parser.add_argument('--download', nargs=2, metavar=('qr', 'name'), help='download a profile using QR code (format: LPA:1$rsp.truphone.com$QRF-SPEEDTEST)')
-  parser.add_argument('--nickname', nargs=2, metavar=('iccid', 'name'), help='update the nickname for a profile')
-  args = parser.parse_args()
-
-  lpa = LPA(interface=args.backend)
-  if args.enable:
-    lpa.enable_profile(args.enable)
-    print('enabled profile, please restart device to apply changes')
-  elif args.disable:
-    lpa.disable_profile(args.disable)
-    print('disabled profile, please restart device to apply changes')
-  elif args.delete:
-    confirm = input('are you sure you want to delete this profile? (y/N) ')
-    if confirm == 'y':
-      lpa.delete_profile(args.delete)
-      print('deleted profile, please restart device to apply changes')
-    else:
-      print('cancelled')
-      exit(0)
-  elif args.download:
-    lpa.download_profile(args.download[0], args.download[1])
-  elif args.nickname:
-    lpa.nickname_profile(args.nickname[0], args.nickname[1])
-  else:
-    parser.print_help()
-
-  profiles = lpa.list_profiles()
-  print(f'\n{len(profiles)} profile{"s" if len(profiles) > 1 else ""}:')
-  for p in profiles:
-    print(f'- {p.iccid} (nickname: {p.nickname or "<none provided>"}) (provider: {p.provider}) - {"enabled" if p.enabled else "disabled"}')
